requirements.txt
asgiref==3.7.2
Brotli==1.0.9
certifi==2023.7.22
cloudinary==1.34.0
dj-database-url==1.2.0
Django==4.2.4
django-cors-headers==3.13.0
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
whitenoise==6.2.0
